requirements.txt
Flask>=2.2.0
PyYAML>=6.0
requests>=2.28.0
wakeonlan>=1.0.0
python-dotenv>=0.21.0
